spring exception handler

+++ spring-learn/src/main/java/com/example/spring_learn/QuickPoll/PollRepository.java
@@ -2,5 +2,5 @@ package com.example.spring_learn.QuickPoll;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface PollRepository extends CrudRepository<Poll, Integer> {
+public interface PollRepository extends CrudRepository<Poll, Long> {
 }
